Mumbai Hacks 2/Mumbai Hacks 2/copilot-money-/backend/requirements.txt
fastapi
uvicorn
python-dotenv
uvicorn
python-dotenv
sqlmodel
alembic
pydantic
groq
yfinance
pandas
httpx
python-multipart
duckduckgo-search
psycopg2-binary
